backend/requirements.txt
Faker==13.11.0
Flask==2.1.2
Flask_Cors==3.0.10
flask_restplus==0.13.0
Flask-RESTful==0.3.9
Flask_SQLAlchemy==2.5.1
Jinja2~=3.0.3
parse==1.19.0
PyJWT>=2.4.0
pytest==7.1.2
python-dotenv==0.20.0
bcrypt==3.2.2
requests>=2.25.1
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug~=2.1.2
flask-socketio
eventlet